добавил создание пользователя

+++ backend/internal/model/habit.go
@@ -3,8 +3,8 @@ package model
 import "time"
 
 type Habit struct {
-	ID          int64     `json:"id"`
-	UserID      int64     `json:"user_id"`
+	ID          int       `json:"id"`
+	UserID      int       `json:"user_id"`
 	Title       string    `json:"title"`
 	Type        string    `json:"type"`
 	TargetValue *int      `json:"target_value,omitempty"`
@@ -12,7 +12,7 @@ type Habit struct {
 }
 
 type HabitInfo struct {
-	UserID      int64  `json:"user_id"`
+	UserID      int    `json:"user_id"`
 	Title       string `json:"title"`
 	Type        string `json:"type"`
 	TargetValue *int   `json:"target_value,omitempty"`
